Implement PartialEq for Line
When `ParialEq` is derived for `Line`, `Line{a, b} != Line{b, a}`. I
think this is wrong as line segments aren't inherently directed even
though it's sometimes convenient to think of them that way. This commit
implements `PartialEq` for `Line` such that `Line{a, b} == Line{b, a}`.

On a practical level, when a line segment is returned as a result of an
intersection operation, the order of the endpoints is an implementation
detail. With derived equality, the intersection of Line `l` with itself
is not always equal to `l`.