Implement total ordering on Point
Total ordering on Point is a requirement of the line sweep algorithm
that will form the basis of further intersection, union, and difference
implementations. It is also very useful for the edge case of line-line
intersection where the two lines are co-linear.

Once the four endpoints of the two lines can be ordered, it's simple to
determine whether the middle two points consitute a gap between the two
lines or the (potentially degenerate) overlap between the two lines.