Add topology-preserving Visvalingam-Whyatt variant
This is based on the technique outlined at
https://www.jasondavies.com/simplify/

It uses an R* tree to restrict the intersection search to only those
segments that fall within a given triangle's bounding box. If an
intersection is detected, the previous point (i.e. the triangle's
"left" member is also marked for removal by setting its area to
negative epsilon), removing the self-intersection.