Correctly check for LineString containment in Polygon
The [DE-9IM specification for [`containment`]
(http://docs.safe.com/fme/html/FME_Desktop_Documentation/FME_Transformers/Transformers/spatialrelations.htm#DE9IM_Matrix) states:

> The **interiors intersect** and no part of the candidate's interior
or boundary intersects the base's exterior. It is possible for the boundaries to intersect.

For `intersection`:

> The two features are not disjoint

For `disjoint`

> The boundaries and interiors do not intersect.

The current containment check deviates from this, in that it forbids
intersection (`!self.intersects(linestring)`), when in fact it should
be forbidding intersection with any _interior rings_. This PR fixes
that, and adds a test.

closes #157