Add validation constructors for LineString, and some other methods.
Add new constructor functions for `LineString`, and a few other misc.
functions:

* `LineString`
  * `LineString::new`
  * `LineString::new_unchecked`
  * `LineString::points`
  * `LineString::points_mut`
  * `LineString::is_closed`
* `LineStringError`
  * `LineStringError::into_vec`

This is backwards compatible such that it doesn't un-`pub`licize the
`LineString` internals (yet). The intention is for this to land, migrate
everything to the new functions, then close off access to the internals.

Partially extracted from https://github.com/georust/rust-geo/pull/155.