Remove duplicate algorithm definitions in 'geo' and 'geo-types'.
To implement Spade’s traits in the geo-types crates, we need to access to a
few geospatial algorithms.

Prior to this commit, the algorithms were just copy/pasted from the
`geo` crate into an `algorithms` module in the `geo-types` crate.

After this commit, we'll have a *hidden, public* `private_utils` module
exposed in `geo-types` so the `geo` crate can reuse these functions and
avoid the duplication.

This whole situation isn't ideal, but I think this is a little better
than duplicating everything.