prefer coord over point for algorithms
we have a pretty-much-only theoretical distinction between the two, but
basically the rule is to prefer Coord over Point for any algorithmic
manipulation.

Long term I think the idea is that a Point might be a top level feature,
with other data, e.g. perhaps with a CRS or something - which wouldn't
apply to this kind of algorithm, which is only concerned with the x/y.