Run CI on pull requests
This ensures CI automatically runs on PRs, even if submitted from forks.

Doing this avoids a silly case of a PR being ready to merge and having a green "merge" button, only for the bors to start test&submitting it, realizing PR has bugs, and waiting for a fix from a user and another review from the maintainer.