Generalize geo-type macros
Make all coordinate-specific logic limited just to the `coord!` macro, and all other macros will pass their parameters as is.

This change is a noop because `coord!` macro still uses the same parameter pattern as before.

This PR removes hardcoded x,y coordinates from the `point!`, `line_string!`, and `polygon!` macros, and instead each macro expects `token: expression` pair(s) that is passed as a parameter to the `coord!` macro.