Allow extra data to be attached to geometries (proof of concept).
For a couple projects I'm working on, I'd like to be able to attach
arbitrary metadata and properties to geometries. Right now that isn't
possible with our geometry types, so this is a proof-of-concept of what
that would look like. Note that this is separate from the XYZ/XYZM work,
as that extends the dimensions of the `Coordinate` type, and this
proof-of-concept does not modify `Coordinate`.