chore: add spade as a dependency
although spade is a optional dependency, it looks like a good fit for
the default features. Including it has the effect of bringing in 3 new
dependencies:

- robust 0.2 (we already use this)
- optional 0.5
- spade 2.2

I'll try to speak to the spade maintainers. Maybe we can bump the robust
version in their crate to reduce dependencies further. Also it might
make sense to ask them if the optional crate provides any real value.