Move total_cmp to HasKernel
The previous approach worked, but then there was feedback that the trait
should maybe be "more complete" and implemented on additional types.

Really, I introduced the trait to solve a problem: we have some
undesirable unwraps in the codebase with some of our floating type
geometries. I didn't care to actually introduce a general-purpose TotalOrd
trait that was useful outside of that context.