fix(review): special case the algo to triangles only
As stated in one of the review comments, the main purpose of the algo is
to stitch together triangles that resulted from a triangulation.

Although it would also be nice to generalize the algo for polygons and
multi polygons, it makes the code harder to read. It is advised to just
triangulate in those cases and then run the stitching on the compound
triangulation.

On another note: special casing the algorithm to triangles also gave us
a real good performance boost again. We gained another ~50% boost from
all of the improvements in this commit.

Authored-by: RobWalt <robwalter96@gmail.com>