Stop using non-existant no `approx` feature
Addresses clippy warning:

> warning: unexpected `cfg` condition value: `approx`

The `geo` crate does not expose a dependency (optional or otherwise) on approx,
though it is used by geo-types,
and there *is* a non-optional dev-dependency on approx,
so I understand the confusion.

Since there is no approx feature in geo, these lines were only triggered
in `cfg(test)` anyway, so lets move them into the test module for simplicity,
which also nicely gets rid of the clippy warning.