Define Length and Densify on metric space
"long ago" `Densify` was defined on the individual geometries: LineString, Line, etc.

Recently, these methods were changed to be generic so they could work
with Euclidean vs. Haversine, etc.

However, to support configurable metric spaces (e.g. Haversine with
custom planet radius) a "type" generic is not enough - we need to pass
a (potentially configured) metric space as an argument.

However, this made the arguments a little awkward:

    line.densify(&Haversine, 100)

Passing `&Haversine` is a little awkward, so this commits switch to
instead be:

    Haversine.densify(&line, 100)

Which allows us to seemlessly add support for something like:

   let mars_radius = 3_389_500
   let mars = CustomHaversine::with_radius(mars_radius);
   mars.densify(&line, 100)