more legible geo-types debug output (WKT-ish)
FIXES #564

Note: Most of the output is valid WKT, but not all of it is.
For example, there is no RECT or TRIANGLE type in standard WKT.

And there are certain invalid representations you can make in geo that
are not valid WKT, but for a debug output format we still need to print
*something*

e.g. a polygon with interiors but no exterior: `POLYGON(EMPTY,(1 2,2 4))`