Fix Relate for Point/MultiPoint
I recently broke this (in 44816ebf, merged yesterday, not released) by
attempting to reuse the bbox from the geometry graph's RTree.

The problem was that the RTree only contains the Geometry's _segments_,
but Point and MultiPoint have no segments - they're zero-dimensional
geometries.

I don't often use `Relate` with zero dimensional geometries, so it's
only luck that an unrelated example I was writing triggered an assert!

== PERF

There was only a small loss from this change, indicating it was probably a bad
idea from the start.

    build prepared polygons time:   [4.5677 ms 4.5717 ms 4.5764 ms]
                            change: [+1.1094% +1.2751% +1.4241%] (p = 0.00 < 0.05)
                            Performance has regressed.

    relate already prepared polygons
                            time:   [18.889 ms 18.912 ms 18.942 ms]
                            change: [+0.5218% +0.6533% +0.8127%] (p = 0.00 < 0.05)
                            Change within noise threshold.

    build and relate prepared polygons
                            time:   [23.598 ms 23.632 ms 23.671 ms]
                            change: [+0.4791% +0.6917% +0.9029%] (p = 0.00 < 0.05)
                            Change within noise threshold.

    unprepared polygons/relate unprepared polygons
                            time:   [744.51 ms 749.08 ms 754.49 ms]
                            change: [-1.2051% -0.1460% +0.8681%] (p = 0.80 > 0.05)
                            No change in performance detected.