Rename features to take advantage of `dep:` syntax
A million years ago, rust introduced the `dep:` syntax to disambiguate
between "the optional crate called serde" and "the feature called
serde", we're finally taking advantage of it here.

I've also attempted to provide some kind of compiler warning to allow
people to migrate to the new feature flag names, since this isn't a
thing the rust compiler natively supports.

Example deprecation output of project using the old `use-rstar` feature:

```
warning: use of deprecated unit struct
`deprecated_feature_flags::UseRstar`: The `use-rstar` feature has been
renamed to simply `rstar`. Use the `rstar` feature instead.
   --> /Users/mkirk/src/georust/geo/geo-types/src/lib.rs:203:35
    |
203 | pub use deprecated_feature_flags::UseRstar;
    |                                   ^^^^^^^^
```

It's not ideal ("unit struct") but I since the compiler doesn't support
deprecating features, this is the best I could come up with.

Note: I did not add an explicit `proj` feature for the deprecated
`use-proj`. Instead we rely on the implicit `proj` feature from the
optional proj crate. I assume the old `use-proj` feature was added as an
act of forward thinking to avoid a breaking change in the case we needed
to enable any accessory crates for the `proj` feature. We can safely do
that now if/when we need to without causing a breaking change.