improve de-9im docs
Render some example geometries in SVG and show the output matrix rendered as a table.

This is the first time we've used SVG in an example. It's a bit verbose,
but I think it renders nicely.

If it's a thing we want to do more of, maybe we can come up with some
machinery to make it easier / DRY.